+++ src/graphql_clj/validator/rules/fields_on_correct_type.clj
@@ -1,0 +1,20 @@
+(ns graphql-clj.validator.rules.fields-on-correct-type
+  "A GraphQL document is only valid if all fields selected are defined by the
+   parent type, or are an allowed meta field such as __typename."
+  (:require [graphql-clj.visitor :refer [defnodevisitor]]
+            [graphql-clj.validator.errors :as ve]
+            [clojure.spec :as s]
+            [graphql-clj.spec :as spec]))
+
+(defn- missing-type-error [{:keys [spec] :as n} s]
+  (format "Cannot query field '%s' on type '%s'." (name spec) (name (spec/get-parent-type n s))))
+
+;; TODO allowed meta field such as __typename?
+
+(defnodevisitor missing-type :pre :field
+  [{:keys [spec v/path] :as n} s]
+  (when-not (s/get-spec spec)
+    {:state (ve/update-errors s (missing-type-error n s))
+     :break true}))
+
+(def rules [missing-type])