+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. This change
 - Implement preliminary version of FieldsOnCorrectType validation
 - Build a map from namespaced spec keyword to their corresponding nodes
 - Include full parent node in each child node (for getting to parent types)
+- Happy path implementation for validating argument and type names.
 
 ## [0.1.16] - 2016-10-12
 ### BREAKING