+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased][unreleased]
+
+## [0.1.17] - 2016-10-14
 - Implement preliminary version of FieldsOnCorrectType validation
 - Build a map from namespaced spec keyword to their corresponding nodes
 - Include full parent node in each child node (for getting to parent types)
 - Validate argument, type, and fragment names
 - Validate that variables are input types
+- Fix required field argument issue
 
 ## [0.1.16] - 2016-10-12
 ### BREAKING
@@ -92,7 +95,8 @@ All notable changes to this project will be documented in this file. This change
 
 ## 0.1.5 - 2016-08-28
 
-[unreleased]: https://github.com/tendant/graphql-clj/compare/0.1.16...HEAD
+[unreleased]: https://github.com/tendant/graphql-clj/compare/0.1.17...HEAD
+[0.1.17]: https://github.com/tendant/graphql-clj/compare/0.1.16...0.1.17
 [0.1.16]: https://github.com/tendant/graphql-clj/compare/0.1.15...0.1.16
 [0.1.15]: https://github.com/tendant/graphql-clj/compare/0.1.14...0.1.15
 [0.1.14]: https://github.com/tendant/graphql-clj/compare/0.1.13...0.1.14